Start implementing Lnd backend service
