--- src/main/Core/Core.ts
@@ -1,84 +1,0 @@
-/* eslint-disable max-classes-per-file */
-import fs from "fs-extra";
-import { spawn } from "child_process";
-import { app, dialog } from "electron";
-import { commandLineArgs } from "_m/common/CommandLineArgs";
-import { RpcServerConfigurations } from "_m/common/RpcServerConfigurations";
-import { Settings } from "_m/Settings/Settings";
-import { Utils } from "../common/Utils";
-
-class Core {
-  private static getDatadir = () => `${Settings.userDataPath()}/Core`;
-
-  private static getArgs = () => {
-    const args = [
-      `--rpcuser=${RpcServerConfigurations.username}`,
-      `--rpcpassword=${RpcServerConfigurations.password}`,
-      `--rpcbind=${RpcServerConfigurations.hostname}`,
-      `--rpcport=${RpcServerConfigurations.port}`,
-      `--zmqpubrawblock=tcp://${RpcServerConfigurations.hostname}:28332`,
-      `--zmqpubrawtx=tcp://${RpcServerConfigurations.hostname}:28333`,
-      "--server=1",
-      "--txindex",
-    ];
-
-    /* istanbul ignore if */
-    if (commandLineArgs.testnet) {
-      args.push("--testnet");
-    }
-
-    if (commandLineArgs.datadir) {
-      args.push(`--datadir=${commandLineArgs.datadir}`);
-    } else {
-      args.push(`--datadir=${Core.getDatadir()}`);
-    }
-
-    return args;
-  };
-
-  private static getPath = () => {
-    const { platform, arch } = Utils.getGlobalProcess();
-    const root = Utils.getAppRoot();
-    const bitcoind =
-      platform === "win32"
-        ? /* istanbul ignore next */ "bitcoind.exe"
-        : "bitcoind";
-
-    return `${root}/bin/${platform}-${arch}/${bitcoind}`;
-  };
-
-  private process: ReturnType<typeof spawn> | null = null;
-
-  spawn = async () => {
-    await fs.ensureDir(Core.getDatadir());
-
-    this.process = spawn(Core.getPath(), Core.getArgs(), {
-      stdio: ["ignore", "ignore", "pipe"],
-    });
-
-    this.process.on(
-      "error",
-      /* istanbul ignore next */ error => {
-        if ((error as any).code === "ENOENT") {
-          dialog.showMessageBoxSync({
-            message: "We're sorry, but your operating system is not supported.",
-          });
-
-          app.exit(0);
-        } else {
-          throw error;
-        }
-      },
-    );
-  };
-
-  stop = () =>
-    new Promise(resolve => {
-      this.process?.on("close", () => {
-        resolve();
-      });
-      this.process?.kill("SIGTERM");
-    });
-}
-
-export const core = new Core();
